skills/wechat-sticker-maker/requirements.txt
Pillow>=9.0.0
rembg>=2.0.0
onnxruntime>=1.14.0
